+++ CHANGELOG.md
@@ -6,11 +6,17 @@
 - CHANGELOG.md!
 - Log viewer / exporter
 - Clone pipeline functionality
+- Added a readonly user and verification for all create/read/delete operations
 
 ### Changed
 - Removed commented out function for simulating pipelines, will implement later
 - Updated license dates
+- Made the showToast function in javascript global and removed duplicated references to it in templates
 
 ### Fixed
 - Missing grok-patterns file, preventing autocomplete in grok debugger from working
-- 
+
+
+
+
+